+++ src/storage/storage_backend.rs
@@ -6,14 +6,14 @@ where
     W: Clone + std::cmp::PartialEq,
 {
     type EdgeIter<'a>: Iterator<Item=Edge<W>> where Self: 'a, W: 'a;
-    fn add_edge_to_node(&mut self, node: u32, edge: &Edge<W>);
+    fn add_edge_to_node(&mut self, node: u64, edge: &Edge<W>);
     fn add_node(&mut self);
-    fn node_len(&self, node: u32) -> usize;
-    fn get_edges<'a>(&'a self, node: u32) -> Self::EdgeIter<'a> where W: 'a;
-    fn remove_edge<F>(&mut self, source: u32, edge: &Edge<W>, func: F) -> Result<Edge<W>, GraphErrors>
+    fn node_len(&self, node: u64) -> usize;
+    fn get_edges<'a>(&'a self, node: u64) -> Self::EdgeIter<'a> where W: 'a;
+    fn remove_edge<F>(&mut self, source: u64, edge: &Edge<W>, func: F) -> Result<Edge<W>, GraphErrors>
     where
         F: Fn(&Edge<W>, &Edge<W>) -> bool;
-    fn contains_edge(&self, source: u32, target: u32) -> Result<Edge<W>, GraphErrors>;
+    fn contains_edge(&self, source: u64, target: u64) -> Result<Edge<W>, GraphErrors>;
     fn node_count(&self) -> usize;
     fn edge_count(&self) -> usize;
     fn increment_node_counter(&mut self);
@@ -21,23 +21,23 @@ where
     // --- Primitives for strategy-driven remove_node ---
 
     /// Clears all outgoing edges from a node and updates the edge count.
-    fn clear_node_edges(&mut self, node: u32);
+    fn clear_node_edges(&mut self, node: u64);
 
     /// Removes the first edge from `source` that points to `target`.
     /// Updates the edge count.
-    fn remove_edge_by_target(&mut self, source: u32, target: u32);
+    fn remove_edge_by_target(&mut self, source: u64, target: u64);
 
     /// Records that `source` has an incoming edge from `origin` (reverse index).
-    fn add_reverse_edge(&mut self, source: u32, origin: u32);
+    fn add_reverse_edge(&mut self, source: u64, origin: u64);
 
     /// Returns all node IDs that have outgoing edges pointing to `node`.
-    fn get_reverse_edges(&self, node: u32) -> Vec<u32>;
+    fn get_reverse_edges(&self, node: u64) -> Vec<u64>;
 
     /// Clears the reverse edge list for a node.
-    fn clear_reverse_edges(&mut self, node: u32);
+    fn clear_reverse_edges(&mut self, node: u64);
 
     /// Removes a single reverse entry: `origin` no longer points to `source`.
-    fn remove_reverse_edge(&mut self, source: u32, origin: u32);
+    fn remove_reverse_edge(&mut self, source: u64, origin: u64);
 
     /// Decrements the node counter.
     fn decrement_node_counter(&mut self);
