+++ src/strategies/direction_strategy.rs
@@ -12,11 +12,11 @@ where
 {
     /// Processes the raw source, target, and weight, mutating the `graph` directly.
     /// Returns the edges that were successfully created.
-    fn add_edge(graph: &mut impl StorageBackend<W>, source: u32, target: u32, weight: &W) -> Result<Edge<W>, GraphErrors>;
+    fn add_edge(graph: &mut impl StorageBackend<W>, source: u64, target: u64, weight: &W) -> Result<Edge<W>, GraphErrors>;
 
-    fn remove_edge(graph: &mut impl StorageBackend<W>, source: u32, target: u32, weight: &W ) -> Result<Edge<W>, GraphErrors>;
+    fn remove_edge(graph: &mut impl StorageBackend<W>, source: u64, target: u64, weight: &W ) -> Result<Edge<W>, GraphErrors>;
 
     /// Removes a node and all edges connected to it.
     /// The strategy determines how to efficiently find and remove incoming edges.
-    fn remove_node(graph: &mut impl StorageBackend<W>, node_id: u32);
+    fn remove_node(graph: &mut impl StorageBackend<W>, node_id: u64);
 }
