+++ src/strategies/weighted.rs
@@ -25,8 +25,8 @@ where
     /// is missing from the graph's adjacency list.
     fn add_edge(
         graph: &mut impl StorageBackend<W>,
-        source: u32, 
-        target: u32, 
+        source: u64, 
+        target: u64, 
         weight: &W
     ) -> Result<Edge<W>, GraphErrors> {
 
@@ -40,7 +40,7 @@ where
         Ok(edge)
     }
 
-    fn remove_edge(graph: &mut impl StorageBackend<W>, source: u32, target: u32, weight: &W) -> Result<Edge<W>, GraphErrors> {
+    fn remove_edge(graph: &mut impl StorageBackend<W>, source: u64, target: u64, weight: &W) -> Result<Edge<W>, GraphErrors> {
         let edge = Edge::new(target, weight);
         let reverse_edge = Edge::new(source, weight);
         graph.remove_edge(target, &reverse_edge, |edge_1: &Edge<W>, edge_2: &Edge<W>| -> bool 
@@ -53,7 +53,7 @@ where
     ///
     /// Since the graph is undirected, every edge in node's outgoing list implies
     /// a reverse edge in the neighbor's list. We use this to avoid a full scan.
-    fn remove_node(graph: &mut impl StorageBackend<W>, node_id: u32) {
+    fn remove_node(graph: &mut impl StorageBackend<W>, node_id: u64) {
         let edges: Vec<Edge<W>> = graph.get_edges(node_id).collect();
         for edge in edges {
             graph.remove_edge_by_target(edge.get_target(), node_id);
